move common test files to common folder
